View README.md for changes

+++ src/sample/SettingsController.java
@@ -6,6 +6,7 @@ import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.*;
+import javafx.util.converter.NumberStringConverter;
 
 import java.net.URL;
 import java.util.ArrayList;
@@ -31,14 +32,25 @@ public class SettingsController implements Initializable {
     private ComboBox profile;
     @FXML
     private Button save;
+    @FXML
+    private Button apply;
+    @FXML
+    private Button delete;
+    @FXML
+    private TextField portText;
 
     private SettingsProfile selectedProfile;
     private List<SettingsBundle> profiles = null;
-    private List<String> profileNames = null;
+    private String startingUrl = null;
+    private String startingPort = null;
+    private String startingClientName = null;
+    private boolean startingCleanSess;
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
-        System.out.println("Initialise called");
+
+        lightThresText.setTextFormatter(new TextFormatter<>(new NumberStringConverter()));
+        proxThresText.setTextFormatter(new TextFormatter<>(new NumberStringConverter()));
 
         DataBaseManager dbManager = new DataBaseManager();
 
@@ -64,9 +76,28 @@ public class SettingsController implements Initializable {
 
         String connUrl = connUrlText.getText();
         String clientName = clientNameText.getText();
+        String port = portText.getText();
         boolean cleanSess = cleanSessCheck.isSelected();
         int lightThres = Integer.parseInt(lightThresText.getText());
+        if (lightThres > 75) {
+            warningLabel.setText("The light threshold cannot be greater than 75%");
+            return;
+        }
         int proxThres = Integer.parseInt(proxThresText.getText());
+        if (proxThres > 5) {
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+            alert.setTitle("Warning!");
+            alert.setHeaderText("Most android phones have a max proximity sensor range of 5." +
+                    "\nSetting the threshold any higher than that could cause the device" +
+                    "\nto continuously send out the warning signal.");
+            alert.setContentText("Are you sure you wish to proceed?");
+            Optional<ButtonType> result = alert.showAndWait();
+            if (result.isPresent()) {
+                if (result.get() == ButtonType.CANCEL) {
+                    return;
+                }
+            }
+        }
 
         if (connUrl.equals("") || clientName.equals("")) {
             warningLabel.setText("Please fill in all of the forms below");
@@ -88,21 +119,36 @@ public class SettingsController implements Initializable {
             }
 
             SettingsBundle bundle = new SettingsBundle(
-                    connUrl, clientName, cleanSess, lightThres, proxThres, dbmanager.getMaxProfileId() + 1
+                    connUrl, port, clientName, cleanSess, lightThres, proxThres, dbmanager.getMaxProfileId() + 1
                     , profName
             );
-            System.out.println("New profile id: " + bundle.getProfId());
             dbmanager.saveNewProfile(bundle);
+            updateProfilesCombo();
         } else {
-            int currId = selectedProfile.getId();
-            String profName = selectedProfile.getProfileName();
+            if (!connUrl.equals(startingUrl)
+                    || !clientName.equals(startingClientName)
+                    || !port.equals(startingPort)
+                    || cleanSess != startingCleanSess) {
+                Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+                alert.setTitle("Warning");
+                alert.setHeaderText("A change you have made requires the MQTT client to reconnect. " +
+                        "This could cause problems if an android phone is connected.");
+                alert.setContentText("Are you sure you wish to continue?");
+                Optional<ButtonType> result = alert.showAndWait();
+                if (result.get() == ButtonType.OK) {
+                    int currId = selectedProfile.getId();
+                    String profName = selectedProfile.getProfileName();
 
-            SettingsBundle bundle = new SettingsBundle(
-                    connUrl, clientName, cleanSess, lightThres, proxThres, currId, profName);
+                    SettingsBundle bundle = new SettingsBundle(
+                            connUrl, port, clientName, cleanSess, lightThres, proxThres, currId, profName);
+                    bundle.print();
+                    dbmanager.updateProfile(bundle);
+                    this.profiles = dbmanager.getAllProfiles();
+                }
+            }
 
-            dbmanager.updateProfile(bundle);
+
         }
-        updateProfilesCombo();
         dbmanager.closeConnection();
     }
 
@@ -115,7 +161,7 @@ public class SettingsController implements Initializable {
 
         dbManager.closeConnection();
 
-        profileNames = new ArrayList<>();
+        List<String> profileNames = new ArrayList<>();
         for (SettingsBundle tmpProfile : profiles) {
             profileNames.add(tmpProfile.getProfName());
         }
@@ -130,14 +176,19 @@ public class SettingsController implements Initializable {
         String profileName = profile.getSelectionModel().getSelectedItem().toString();
         for (SettingsBundle tmpProfile : profiles) {
             if (profileName.equals(tmpProfile.getProfName())) {
-                System.out.println("Deleting profile: " + tmpProfile.getProfName());
-                DataBaseManager dbManager = new DataBaseManager();
-                SettingsProfile delProfile = new SettingsProfile(tmpProfile.getProfId()
-                        , tmpProfile.getProfName());
-                dbManager.deleteProfile(delProfile);
-                int selectedItem = profile.getSelectionModel().getSelectedIndex();
-                profile.getSelectionModel().clearAndSelect(selectedItem - 1);
-                updateProfilesCombo();
+                if (profileName.equals(Constants.DEFAULT_PROFILE)) {
+                    warningLabel.setText("You cannot delete the default profile");
+                } else {
+                    warningLabel.setText("");
+                    System.out.println("Deleting profile: " + tmpProfile.getProfName());
+                    DataBaseManager dbManager = new DataBaseManager();
+                    SettingsProfile delProfile = new SettingsProfile(tmpProfile.getProfId()
+                            , tmpProfile.getProfName());
+                    dbManager.deleteProfile(delProfile);
+                    int selectedItem = profile.getSelectionModel().getSelectedIndex();
+                    profile.getSelectionModel().clearAndSelect(selectedItem - 1);
+                    updateProfilesCombo();
+                }
             }
         }
     }
@@ -145,16 +196,39 @@ public class SettingsController implements Initializable {
     @FXML
     private void handleComboBoxAction() {
         if (profile != null) {
-            String profileName = profile.getSelectionModel().getSelectedItem().toString();
-            for (SettingsBundle tmpProfile : profiles) {
-                if (profileName.equals(tmpProfile.getProfName())) {
-                    connUrlText.setText(tmpProfile.getConnUrl());
-                    clientNameText.setText(tmpProfile.getClientName());
-                    cleanSessCheck.setSelected(tmpProfile.getCleanSess());
-                    lightThresText.setText(String.valueOf(tmpProfile.getLightThres()));
-                    proxThresText.setText(String.valueOf(tmpProfile.getProxThres()));
+            if (profiles != null) {
+                String profileName = profile.getSelectionModel().getSelectedItem().toString();
+                for (SettingsBundle tmpProfile : profiles) {
+                    if (profileName.equals(tmpProfile.getProfName())) {
+                        connUrlText.setText(tmpProfile.getConnUrl());
+                        startingUrl = tmpProfile.getConnUrl();
+                        portText.setText(tmpProfile.getPort());
+                        startingPort = tmpProfile.getPort();
+                        clientNameText.setText(tmpProfile.getClientName());
+                        startingClientName = tmpProfile.getClientName();
+                        cleanSessCheck.setSelected(tmpProfile.getCleanSess());
+                        startingCleanSess = tmpProfile.getCleanSess();
+                        lightThresText.setText(String.valueOf(tmpProfile.getLightThres()));
+                        proxThresText.setText(String.valueOf(tmpProfile.getProxThres()));
+                        DataBaseManager dbManager = new DataBaseManager();
+                        dbManager.switchProfile(tmpProfile.getProfId());
+                    }
                 }
             }
         }
     }
+
+    @FXML
+    private void filterText(String text) {
+        if (!text.matches("\\d*")) {
+
+        }
+    }
+
+    @FXML
+    private void apply() {
+        save();
+        Main.settingsChanged();
+    }
+
 }
